+++ src/cpp/encoder/encoder.hpp
@@ -6,15 +6,15 @@
 class PoissonEncoder
 {
   int obs_dim;
-  float max_freq;
+  double max_freq;
 
   std::mt19937 mt;
-  std::uniform_real_distribution<float> rand_01;
+  std::uniform_real_distribution<double> rand_01;
 
 public:
-  PoissonEncoder(int obs_dim, float max_freq);
+  PoissonEncoder(int obs_dim, double max_freq);
 
-  std::vector<float> encode(const std::vector<float>& obs, float dt);
+  std::vector<double> encode(const std::vector<double>& obs, double dt);
 };
 
 class PopSANEncoder
@@ -22,53 +22,53 @@ class PopSANEncoder
   int obs_dim;
   int pop_dim;
   int spike_ts;
-  float std;
-  float v_reset;
-  float v_th;
+  double std;
+  double v_reset;
+  double v_th;
 
   int encoder_neuron_dim;
-  std::vector<float> mean;
+  std::vector<double> mean;
 
 public:
   PopSANEncoder(
     int obs_dim,
     int pop_dim,
     int spike_ts,
-    float std,
-    float v_reset,
-    float v_th,
-    const std::pair<float, float>& mean_range);
+    double std,
+    double v_reset,
+    double v_th,
+    const std::pair<double, double>& mean_range);
 
-  std::vector<std::vector<float>> encode(const std::vector<float>& obs);
+  std::vector<std::vector<double>> encode(const std::vector<double>& obs);
 };
 
 class LIFEncoder
 {
   int obs_dim;
   int pop_dim;
-  float std;
-  float v_rest;
-  float v_reset;
-  float v_th;
-  float tau_m;
-  float amp;
+  double std;
+  double v_rest;
+  double v_reset;
+  double v_th;
+  double tau_m;
+  double amp;
 
   int encoder_neuron_dim;
-  std::vector<float> mean, pop_volt, pop_spikes;
+  std::vector<double> mean, pop_volt, pop_spikes;
 
 public:
   LIFEncoder(
     int obs_dim,
     int pop_dim,
-    const std::pair<float, float>& mean_range,
-    float std,
-    float v_rest,
-    float v_reset,
-    float v_th,
-    float tau_m,
-    float amp);
+    const std::pair<double, double>& mean_range,
+    double std,
+    double v_rest,
+    double v_reset,
+    double v_th,
+    double tau_m,
+    double amp);
 
-  std::vector<float> encode(const std::vector<float>& obs, float dt);
+  std::vector<double> encode(const std::vector<double>& obs, double dt);
 
-  void reset(const std::vector<float>& init_v);
+  void reset(const std::vector<double>& init_v);
 };
